requirements.txt
requests
selenium==4.9.1
amazoncaptcha
PyAutoGUI
pandas
google-api-core==2.11.0
google-api-python-client==2.77.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
gspread==5.7.2
pyclick
scrapingbee
emoji
beautifulsoup4
opencv-python
loguru
pandas-gbq==0.19.2
python-dotenv